requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
PyYAML==6.0.1
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
httpx==0.25.2
openai==1.3.7
anthropic==0.7.7
langchain==0.0.340
langchain-openai==0.0.2
gitpython==3.1.40
PyGithub==1.59.1
docker==6.1.3
psutil==5.9.6
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
celery==5.3.4
websockets==12.0
streamlit==1.28.2
plotly==5.17.0
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
bandit==1.7.5
safety==2.3.5
pylint==3.0.3
black==23.11.0
mypy==1.7.1
coverage==7.3.2
pre-commit==3.6.0
rich==13.7.0
typer==0.9.0
jinja2==3.1.2
watchdog==3.0.0
